flaskApi/jogoTeca/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
Werkzeug==2.2.2
blinker==1.8.2
click==8.1.7
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
protobuf==5.28.2
Werkzeug==2.2.2
blinker==1.8.2
click==8.1.7
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mysql-connector-python==8.0.28
protobuf==5.28.2
Werkzeug==2.2.2
